+++ .github/workflows/telegram-notify.yml
@@ -1,0 +1,28 @@
+name: Telegram Notification on Push
+
+on:
+  push:
+    branches:
+      - '**'  # This will trigger on pushes to all branches
+
+jobs:
+  notify:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Send Telegram Notification
+      uses: appleboy/telegram-action@master
+      env:
+        GITHUB_CONTEXT: ${{ toJson(github) }}
+      with:
+          to: ${{ secrets.TELEGRAM_CHAT_ID }}
+          token: ${{ secrets.TELEGRAM_BOT_TOKEN }}
+          format: markdown
+          message: |
+            *${{ github.event.head_commit.author.username }}* pushed a new commit to *${{ github.repository }}*.
+            Branch: *${{ github.ref_name }}*
+            Message: 
+            ${{ github.event.head_commit.message }}
+
+            See changes: https://github.com/${{ github.repository }}/commit/${{github.sha}}
+            Timestamp: ${{ github.event.head_commit.timestamp }}
